Move `safe_call` to Utils
